+++ src/components/MyProfile/MyProfile.module.css
@@ -1,0 +1,9 @@
+.div {
+  margin: 2rem;
+}
+
+.div li {
+  padding: 2rem;
+  border: solid black;
+  border-width: 1px;
+}
