+++ src/components/Navbar.js
@@ -1,11 +1,12 @@
 import React from 'react';
 import { NavLink } from 'react-router-dom';
 import planet from './planet.png';
+import classes from './Navbar.module.css';
 
 const Navigation = () => (
   <div className="nav-bar">
     <div className="inner-nav-bar">
-      <div className="logo-wrapper">
+      <div className={classes.logo}>
         <img src={planet} alt="id" width="40px" height="40px" />
         <h1>Space Traveler&#39;s Hub</h1>
       </div>
